changed double money values to BigDecimal

+++ src/main/java/ru/j3v/broker/ExchangeService.java
@@ -4,6 +4,7 @@ import ru.j3v.io.DataReader;
 import ru.j3v.io.FileDataReader;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.util.*;
 
 @Service
@@ -11,8 +12,8 @@ public class ExchangeService {
 
     private TimeService timeService;
 
-    private Map<String, Map<Date, Double>> assetsPrices;
-    private Map<String, Map<Date, Double>> currencyPrices;
+    private Map<String, Map<Date, BigDecimal>> assetsPrices;
+    private Map<String, Map<Date, BigDecimal>> currencyPrices;
     private Map<String, String> assetCurrency;
 
     public ExchangeService(TimeService timeService) {
@@ -28,7 +29,7 @@ public class ExchangeService {
         return assetCurrency.get(asset);
     }
 
-    public double getPrice(String asset) {
+    public BigDecimal getPrice(String asset) {
         Date date = timeService.getCurrentDate();
         return assetsPrices.get(asset).get(date);
     }
